pyproject.toml
[project]
name = "django-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.4",
    "djangorestframework>=3.16.0",
    "mypy>=1.17.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.3",
]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = "3.12"

[tool.django-stubs]
django_settings_module = "Django_learning.settings"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["Django_learning", "projects"]
